+++ src/potent/commands/init.py
@@ -3,8 +3,8 @@ from typing import Annotated
 
 from cyclopts import App, Parameter, validators
 
-from potent.commands._types import is_plan_json, pathify
-from potent.plan import Plan
+from potent.commands._types import get_command_dir, is_plan_json, pathify
+from potent.plan import CommandConfig, Plan, PlanConfig
 
 app = App()
 
@@ -25,9 +25,17 @@ def init(
     /,
 ):
     """
-    Create an empty plan at the specified path.
+    Create an empty plan at the specified path. If the path resolves to the config directory, then it defaults to `command` mode. Otherwise, the default of `plan` is used.
     """
 
+    is_command = get_command_dir() in path.parents
+
+    path.parent.mkdir(exist_ok=True, parents=True)
     path.write_text(
-        Plan(version="v1", operations=[], directories=[]).model_dump_json(indent=2)
+        Plan(
+            version="v1",
+            operations=[],
+            directories=[],
+            config=CommandConfig() if is_command else PlanConfig(),
+        ).model_dump_json(indent=2)
     )